upgrade to latest serenity-next

+++ botv2/src/ipc/argparse.rs
@@ -15,12 +15,12 @@
      */
 #[derive(Debug, Clone)]
 pub struct MewldCmdArgs {
-    pub shards: Vec<u32>,
-    pub shard_count: u32,
-    pub cluster_id: u32,
+    pub shards: Vec<u16>,
+    pub shard_count: u16,
+    pub cluster_id: u16,
     pub cluster_name: String,
     pub base_dir: String,
-    pub cluster_count: u32,
+    pub cluster_count: u16,
     pub splashtail_url: String,
     pub mewld_redis_channel: String,
 }
@@ -34,12 +34,12 @@ Expected arguments: [program name] <shards> <shard_count> <cluster_id> <cluster_
             "#.into());
         }
 
-        let shards: Vec<u32> = serde_json::from_str(&args[1])?;
-        let shard_count: u32 = str::parse(&args[2])?;
-        let cluster_id: u32 = str::parse(&args[3])?;
+        let shards: Vec<u16> = serde_json::from_str(&args[1])?;
+        let shard_count: u16 = str::parse(&args[2])?;
+        let cluster_id: u16 = str::parse(&args[3])?;
         let cluster_name: String = args[4].clone();
         let base_dir: String = args[5].clone();
-        let cluster_count: u32 = str::parse(&args[6])?;
+        let cluster_count: u16 = str::parse(&args[6])?;
         let splashtail_url: String = args[7].clone();
         let mewld_redis_channel: String = args[8].clone();
